swap the old and new debugger source files
