+++ styles/debugger.less
@@ -128,18 +128,19 @@
 }
 
 .debugger-section.resizable {
-  min-width: 160px;
+  min-width: 100px;
   min-height: 100px;
 }
 
 .debugger {
   .debugger-execution-toolbar {
     display: flex;
-    margin-right: -3px;
+    margin-left: -3px;
+    overflow-x: hidden;
 
     button.btn {
-      width: 40px;
-      margin-right: 3px;
+      // width: 40px;
+      margin-left: 3px;
     }
   }
 }
