+++ styles/views.less
@@ -86,6 +86,10 @@ div.atom-view {
     .view-subtitle {
       margin-top: 2px;
       color: @text-color-subtle;
+      // TODO: Overflow or not?
+      overflow: hidden;
+      white-space: nowrap;
+      text-overflow: ellipsis;
     }
   }
